Add Google+ authentication

+++ restaurant.py
@@ -1,17 +1,32 @@
+import httplib2
+import json
 import os
+import random
+import requests
+import string
 
 from configparser import ConfigParser
 from flask import (flash,
                    Flask,
                    jsonify,
+                   make_response,
                    redirect,
                    render_template,
                    request,
+                   session as login_session,
                    url_for)
+from oauth2client.client import flow_from_clientsecrets # for GConnect
+from oauth2client.client import FlowExchangeError # for GConnect
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem
 
+
+GOOGLE_JSON_FILE = 'client_secret.json'
+
+CLIENT_ID = json.loads(
+    open(GOOGLE_JSON_FILE, 'r').read())['web']['client_id']
+APPLICATION_NAME = "restaurant_app"
 
 config = ConfigParser()
 config.read('config.ini')
@@ -26,6 +41,130 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
+
+# Create Anti-Forgery State Token
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in range(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Obtain authorization code
+    code = request.data
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets(GOOGLE_JSON_FILE, scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+        json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1].decode('utf8'))
+
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print("Token's client ID does not match app's.")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check to see if user is already logged in
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info with Google API
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+    data = answer.json()
+
+    # Store the login session
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    output = []
+    output.append('<h1>Welcome, ')
+    output.append(login_session['username'])
+    output.append('!</h1>')
+    output.append('<img src="')
+    output.append(login_session['picture'])
+    output.append(' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> ')
+    flash("you are now logged in as %s" % login_session['username'])
+    return ''.join(output)
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        print('Access Token is None')
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    print('In gdisconnect access token is %s', access_token)
+    print('User name is: ')
+    print(login_session['username'])
+    url = ('https://accounts.google.com/o/oauth2/revoke?token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print('result is ')
+    print(result)
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 @app.route('/')
 @app.route('/restaurants')
 def showRestaurants():
@@ -34,6 +173,8 @@ def showRestaurants():
 
 @app.route('/restaurant/new', methods=['GET', 'POST'])
 def newRestaurant():
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         new_restaurant = Restaurant(name=request.form['name'],
                                     description=request.form['description'])
@@ -46,6 +187,8 @@ def newRestaurant():
 
 @app.route('/restaurant/<int:restaurant_id>/edit', methods=['GET', 'POST'])
 def editRestaurant(restaurant_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     edited_restaurant = session.query(Restaurant).filter_by(id=restaurant_id) \
                                                  .one()
     if request.method == 'POST':
@@ -68,16 +211,18 @@ def editRestaurant(restaurant_id):
 
 @app.route('/restaurant/<int:restaurant_id>/delete', methods=['GET', 'POST'])
 def deleteRestaurant(restaurant_id):
-     restaurant_to_delete = session.query(Restaurant) \
-                                   .filter_by(id=restaurant_id).one()
-     if request.method == 'POST':
-         session.delete(restaurant_to_delete)
-         session.commit()
-         flash("Restaurant has been deleted")
-         return redirect(url_for('showRestaurants'))
-     else:
-         return render_template('deleterestaurant.html',
-                                restaurant=restaurant_to_delete)
+    if 'username' not in login_session:
+        return redirect('/login')
+    restaurant_to_delete = session.query(Restaurant) \
+                                  .filter_by(id=restaurant_id).one()
+    if request.method == 'POST':
+        session.delete(restaurant_to_delete)
+        session.commit()
+        flash("Restaurant has been deleted")
+        return redirect(url_for('showRestaurants'))
+    else:
+        return render_template('deleterestaurant.html',
+                               restaurant=restaurant_to_delete)
 
 @app.route('/restaurants/JSON')
 def showRestaurantsJSON():
@@ -85,6 +230,7 @@ def showRestaurantsJSON():
     return jsonify(Restaurants=[restaurant.serialize
                                 for restaurant in restaurants])
 
+#JSON APIs to view Restaurant Information
 @app.route('/restaurants/<int:restaurant_id>/menu/JSON')
 def restaurantMenuJSON(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
@@ -92,7 +238,6 @@ def restaurantMenuJSON(restaurant_id):
         restaurant_id=restaurant_id).all()
     return jsonify(MenuItems=[i.serialize for i in items])
 
-# JSON ENDPOINT
 @app.route('/restaurants/<int:restaurant_id>/menu/<int:menu_id>/JSON')
 def menuItemJSON(restaurant_id, menu_id):
     menu_item = session.query(MenuItem).filter_by(id=menu_id).one()
@@ -107,6 +252,8 @@ def restaurantMenu(restaurant_id):
 
 @app.route('/restaurants/<int:restaurant_id>/new', methods=['GET', 'POST'])
 def newMenuItem(restaurant_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         new_item = MenuItem(name=request.form['name'],
                             description=request.form['description'],
@@ -123,6 +270,8 @@ def newMenuItem(restaurant_id):
 @app.route('/restaurants/<int:restaurant_id>/<int:menu_id>/edit',
            methods=['GET', 'POST'])
 def editMenuItem(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     edited_item = session.query(MenuItem).filter_by(id=menu_id).one()
     if request.method == 'POST':
         if request.form['name']:
@@ -140,6 +289,8 @@ def editMenuItem(restaurant_id, menu_id):
 @app.route('/restaurants/<int:restaurant_id>/<int:menu_id>/delete',
            methods=['GET', 'POST'])
 def deleteMenuItem(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     item_to_delete = session.query(MenuItem).filter_by(id=menu_id).one()
     if request.method == 'POST':
         session.delete(item_to_delete)
